requirements.txt
# ClaimLens Dependencies
# Python 3.11+ required

# Core Framework
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0

# LLM Provider
openai>=1.0.0

# Web Search APIs
tavily-python>=0.3.0
google-search-results>=2.4.0  # SerpAPI

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# NLI Models (HuggingFace)
transformers>=4.36.0
torch>=2.1.0

# Async Support
httpx>=0.26.0
aiohttp>=3.9.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
black>=23.12.0
isort>=5.13.0
mypy>=1.8.0
